+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from './ng_module';
 export * from './storage';