+++ src/ng_module.ts
@@ -1,7 +1,0 @@
-import {NgModule} from '@angular/core';
-
-/**
- * The Web Storage module.
- */
-@NgModule()
-export class NgxWebStorage {}