+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Breaking change: changed the case of the injection tokens.
 - Breaking change: dropped the [CommonJS](https://nodejs.org/api/modules.html) modules in favor of [ECMAScript](http://www.ecma-international.org/ecma-262/6.0/#sec-modules) ones.
 - Breaking change: ported the source code to [TypeScript](https://www.typescriptlang.org).
-- Breaking change: renamed the `StorageModule` module to `NgxWebStorage`.
+- Breaking change: removed the `StorageModule` module.
 - Added a user guide based on [MkDocs](http://www.mkdocs.org).
 - Replaced [ESDoc](https://esdoc.org) documentation generator by [TypeDoc](https://typedoc.org).
 - Replaced [ESLint](https://eslint.org) static analyzer by [TSLint](https://palantir.github.io/tslint).